editor-platform/services/editor-api/pyproject.toml
[project]
name = "editor-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "strawberry-graphql[fastapi]>=0.235",
  "SQLAlchemy>=2.0",
  "asyncpg>=0.29",
  "python-dotenv>=1.0",
  "bleach>=6.1",
  "markdown-it-py>=3.0",
  "beautifulsoup4>=4.12",
]

[tool.uvicorn]
factory = false
